+++ src/SignaturePad.Forms.Platform.Shared/SignaturePad.Forms.Platform.Shared.projitems
@@ -1,0 +1,15 @@
+﻿<?xml version="1.0" encoding="utf-8"?>
+<Project xmlns="http://schemas.microsoft.com/developer/msbuild/2003">
+  <PropertyGroup>
+    <MSBuildAllProjects>$(MSBuildAllProjects);$(MSBuildThisFileFullPath)</MSBuildAllProjects>
+    <HasSharedItems>true</HasSharedItems>
+    <SharedGUID>4c55bdd7-f049-4742-9e00-cdd480e53a54</SharedGUID>
+  </PropertyGroup>
+  <PropertyGroup Label="Configuration">
+    <Import_RootNamespace>SignaturePad.Forms.Platform.Shared</Import_RootNamespace>
+  </PropertyGroup>
+  <ItemGroup>
+    <Compile Include="$(MSBuildThisFileDirectory)ColorExtensions.cs" />
+    <Compile Include="$(MSBuildThisFileDirectory)SignaturePadRenderer.cs" />
+  </ItemGroup>
+</Project>
