+++ src/SignaturePad.Forms.Platform.Shared/SignaturePadRenderer.cs
@@ -1,20 +1,47 @@
 using System.ComponentModel;
 using System.IO;
-using System.Linq;
 using System.Threading.Tasks;
-using Android.Graphics;
+using System.Linq;
 using Xamarin.Forms;
-using Xamarin.Forms.Platform.Android;
 using SignaturePad.Forms;
+using SignaturePad.Forms.Platform;
+using Color = Xamarin.Forms.Color;
+using Point = Xamarin.Forms.Point;
+#if WINDOWS_PHONE
+using System.Windows.Media;
+using System.Windows.Media.Imaging;
+using ImageTools;
+using ImageTools.IO.Png;
+using Xamarin.Forms.Platform.WinPhone;
+using SignaturePad.Forms.WindowsPhone;
+using NativeSignaturePadView = Xamarin.Controls.SignaturePad;
+using NativePoint = System.Windows.Point;
+#elif __IOS__
+using UIKit;
+using Xamarin.Forms.Platform.iOS;
+using SignaturePad.Forms.iOS;
+using NativeSignaturePadView = SignaturePad.SignaturePadView;
+using NativePoint = CoreGraphics.CGPoint;
+using NativeColor = UIKit.UIColor;
+#elif __ANDROID__
+using Android.Graphics;
+using Android.Widget;
+using Xamarin.Forms.Platform.Android;
 using SignaturePad.Forms.Droid;
 using NativeSignaturePadView = SignaturePad.SignaturePadView;
 using NativePoint = System.Drawing.PointF;
-using Color = Xamarin.Forms.Color;
-using Point = Xamarin.Forms.Point;
+using NativeColor = Android.Graphics.Color;
+#endif
 
 [assembly: ExportRenderer(typeof(SignaturePadView), typeof(SignaturePadRenderer))]
 
+#if WINDOWS_PHONE
+namespace SignaturePad.Forms.WindowsPhone
+#elif __IOS__
+namespace SignaturePad.Forms.iOS
+#elif __ANDROID__
 namespace SignaturePad.Forms.Droid
+#endif
 {
     public class SignaturePadRenderer : ViewRenderer<SignaturePadView, NativeSignaturePadView>
     {
@@ -25,7 +52,11 @@ namespace SignaturePad.Forms.Droid
             if (Control == null)
             {
                 // Instantiate the native control and assign it to the Control property
+#if __ANDROID__
                 var native = new NativeSignaturePadView(Xamarin.Forms.Forms.Context);
+#else
+                var native = new NativeSignaturePadView();
+#endif
                 SetNativeControl(native);
             }
 
@@ -63,6 +94,42 @@ namespace SignaturePad.Forms.Droid
             if (ctrl != null)
             {
                 var image = ctrl.GetImage();
+#if WINDOWS_PHONE
+                ExtendedImage img = null;
+                if (e.ImageFormat == SignatureImageFormat.Png)
+                {
+                    img = image.ToImage();
+                }
+                var stream = new MemoryStream();
+                e.ImageStreamTask = Task.Run<Stream>(() =>
+                {
+                    if (e.ImageFormat == SignatureImageFormat.Png)
+                    {
+                        var encoder = new PngEncoder();
+                        encoder.Encode(img, stream);
+                        return stream;
+                    }
+                    if (e.ImageFormat == SignatureImageFormat.Jpg)
+                    {
+                        image.SaveJpeg(stream, image.PixelWidth, image.PixelHeight, 0, 100);
+                        return stream;
+                    }
+                    return null;
+                });
+#elif __IOS__
+                e.ImageStreamTask = Task.Run(() =>
+                {
+                    if (e.ImageFormat == SignatureImageFormat.Png)
+                    {
+                        return image.AsPNG().AsStream();
+                    }
+                    if (e.ImageFormat == SignatureImageFormat.Jpg)
+                    {
+                        return image.AsJPEG().AsStream();
+                    }
+                    return null;
+                });
+#elif __ANDROID__
                 var stream = new MemoryStream();
                 var format = e.ImageFormat == SignatureImageFormat.Png ? Bitmap.CompressFormat.Png : Bitmap.CompressFormat.Jpeg;
                 e.ImageStreamTask = image
@@ -81,6 +148,7 @@ namespace SignaturePad.Forms.Droid
                             return null;
                         }
                     });
+#endif
             }
         }
 
@@ -123,7 +191,7 @@ namespace SignaturePad.Forms.Droid
 
             if (Element.BackgroundColor != Color.Default)
             {
-                Control.BackgroundColor = Element.BackgroundColor.ToAndroid();
+                Control.BackgroundColor = Element.BackgroundColor.ToNative();
             }
             if (!string.IsNullOrEmpty(Element.CaptionText))
             {
@@ -131,7 +199,7 @@ namespace SignaturePad.Forms.Droid
             }
             if (Element.CaptionTextColor != Color.Default)
             {
-                Control.Caption.SetTextColor(Element.CaptionTextColor.ToAndroid());
+                Control.Caption.SetTextColor(Element.CaptionTextColor);
             }
             if (!string.IsNullOrEmpty(Element.ClearText))
             {
@@ -139,7 +207,7 @@ namespace SignaturePad.Forms.Droid
             }
             if (Element.ClearTextColor != Color.Default)
             {
-                Control.ClearLabel.SetTextColor(Element.ClearTextColor.ToAndroid());
+                Control.ClearLabel.SetTextColor(Element.ClearTextColor);
             }
             if (!string.IsNullOrEmpty(Element.PromptText))
             {
@@ -147,15 +215,20 @@ namespace SignaturePad.Forms.Droid
             }
             if (Element.PromptTextColor != Color.Default)
             {
-                Control.SignaturePrompt.SetTextColor(Element.PromptTextColor.ToAndroid());
+                Control.SignaturePrompt.SetTextColor(Element.PromptTextColor);
             }
             if (Element.SignatureLineColor != Color.Default)
             {
-                Control.SignatureLineColor = Element.SignatureLineColor.ToAndroid();
+                var color = Element.SignatureLineColor.ToNative();
+#if WINDOWS_PHONE
+                Control.SignatureLineBrush = new SolidColorBrush(color);
+#else
+                Control.SignatureLineColor = color;
+#endif
             }
             if (Element.StrokeColor != Color.Default)
             {
-                Control.StrokeColor = Element.StrokeColor.ToAndroid();
+                Control.StrokeColor = Element.StrokeColor.ToNative();
             }
             if (Element.StrokeWidth > 0)
             {
@@ -175,43 +248,48 @@ namespace SignaturePad.Forms.Droid
 
             if (property == SignaturePadView.BackgroundColorProperty.PropertyName)
             {
-                this.Control.BackgroundColor = Element.BackgroundColor.ToAndroid();
+                Control.BackgroundColor = Element.BackgroundColor.ToNative();
             }
             else if (property == SignaturePadView.CaptionTextProperty.PropertyName)
             {
-                this.Control.CaptionText = Element.CaptionText;
+                Control.CaptionText = Element.CaptionText;
             }
             else if (property == SignaturePadView.CaptionTextColorProperty.PropertyName)
             {
-                this.Control.Caption.SetTextColor(Element.CaptionTextColor.ToAndroid());
+                Control.Caption.SetTextColor(Element.CaptionTextColor);
             }
             else if (property == SignaturePadView.ClearTextProperty.PropertyName)
             {
-                this.Control.ClearLabelText = Element.ClearText;
+                Control.ClearLabelText = Element.ClearText;
             }
             else if (property == SignaturePadView.ClearTextColorProperty.PropertyName)
             {
-                this.Control.ClearLabel.SetTextColor(Element.ClearTextColor.ToAndroid());
+                Control.ClearLabel.SetTextColor(Element.ClearTextColor);
             }
             else if (property == SignaturePadView.PromptTextProperty.PropertyName)
             {
-                this.Control.SignaturePromptText = Element.PromptText;
+                Control.SignaturePromptText = Element.PromptText;
             }
             else if (property == SignaturePadView.PromptTextColorProperty.PropertyName)
             {
-                this.Control.SignaturePrompt.SetTextColor(Element.PromptTextColor.ToAndroid());
+                Control.SignaturePrompt.SetTextColor(Element.PromptTextColor);
             }
             else if (property == SignaturePadView.SignatureLineColorProperty.PropertyName)
             {
-                this.Control.SignatureLineColor = Element.SignatureLineColor.ToAndroid();
+                var color = Element.SignatureLineColor.ToNative();
+#if WINDOWS_PHONE
+                Control.SignatureLineBrush = new SolidColorBrush(color);
+#else
+                Control.SignatureLineColor = color;
+#endif
             }
             else if (property == SignaturePadView.StrokeColorProperty.PropertyName)
             {
-                this.Control.StrokeColor = Element.StrokeColor.ToAndroid();
+                Control.StrokeColor = Element.StrokeColor.ToNative();
             }
             else if (property == SignaturePadView.StrokeWidthProperty.PropertyName)
             {
-                this.Control.StrokeWidth = Element.StrokeWidth;
+                Control.StrokeWidth = Element.StrokeWidth;
             }
         }
     }
