+++ src/SignaturePad.Forms.WindowsPhone/SignaturePad.Forms.WindowsPhone.csproj
@@ -97,7 +97,6 @@
       <Link>Properties\AssemblyInfo.Forms.cs</Link>
     </Compile>
     <Compile Include="Properties\AssemblyInfo.cs" />
-    <Compile Include="SignaturePadRenderer.cs" />
   </ItemGroup>
   <ItemGroup>
     <ProjectReference Include="..\SignaturePad.Forms\SignaturePad.Forms.csproj">
@@ -171,6 +170,7 @@
     <None Include="packages.config" />
   </ItemGroup>
   <ItemGroup />
+  <Import Project="..\SignaturePad.Forms.Platform.Shared\SignaturePad.Forms.Platform.Shared.projitems" Label="Shared" />
   <Import Project="$(MSBuildExtensionsPath)\Microsoft\$(TargetFrameworkIdentifier)\$(TargetFrameworkVersion)\Microsoft.$(TargetFrameworkIdentifier).$(TargetFrameworkVersion).Overrides.targets" />
   <Import Project="$(MSBuildExtensionsPath)\Microsoft\$(TargetFrameworkIdentifier)\$(TargetFrameworkVersion)\Microsoft.$(TargetFrameworkIdentifier).CSharp.targets" />
   <ProjectExtensions />
