--- src/SignaturePad.Forms.iOS/SignaturePadRenderer.cs
@@ -1,215 +1,0 @@
-using System;
-using System.ComponentModel;
-using CoreGraphics;
-using System.Linq;
-using SignaturePad.Forms;
-using SignaturePad.Forms.iOS;
-using UIKit;
-using Xamarin.Forms;
-using Xamarin.Forms.Platform;
-using Xamarin.Forms.Platform.iOS;
-using NativeSignaturePadView = SignaturePad.SignaturePadView;
-using NativePoint = CoreGraphics.CGPoint;
-using Color = Xamarin.Forms.Color;
-using Point = Xamarin.Forms.Point;
-using System.IO;
-using System.Threading.Tasks;
-
-[assembly: ExportRenderer(typeof(SignaturePadView), typeof(SignaturePadRenderer))]
-
-namespace SignaturePad.Forms.iOS
-{
-    public class SignaturePadRenderer : ViewRenderer<SignaturePadView, NativeSignaturePadView>
-    {
-        protected override void OnElementChanged(ElementChangedEventArgs<SignaturePadView> e)
-        {
-            base.OnElementChanged(e);
-
-            if (Control == null)
-            {
-                // Instantiate the native control and assign it to the Control property
-                var native = new NativeSignaturePadView();
-                SetNativeControl(native);
-            }
-
-            if (e.OldElement != null)
-            {
-                // Unsubscribe from event handlers and cleanup any resources
-                e.OldElement.ImageStreamRequested -= OnImageStreamRequested;
-                e.OldElement.IsBlankRequested -= OnIsBlankRequested;
-                e.OldElement.PointsRequested -= OnPointsRequested;
-                e.OldElement.PointsSpecified -= OnPointsSpecified;
-            }
-
-            if (e.NewElement != null)
-            {
-                // Configure the control and subscribe to event handlers
-                e.NewElement.ImageStreamRequested += OnImageStreamRequested;
-                e.NewElement.IsBlankRequested += OnIsBlankRequested;
-                e.NewElement.PointsRequested += OnPointsRequested;
-                e.NewElement.PointsSpecified += OnPointsSpecified;
-
-                UpdateAll();
-            }
-        }
-
-        protected override void OnElementPropertyChanged(object sender, PropertyChangedEventArgs e)
-        {
-            base.OnElementPropertyChanged(sender, e);
-
-            Update(e.PropertyName);
-        }
-
-        private void OnImageStreamRequested(object sender, SignaturePadView.ImageStreamRequestedEventArgs e)
-        {
-            var ctrl = Control;
-            if (ctrl != null)
-            {
-                var image = ctrl.GetImage();
-                e.ImageStreamTask = Task.Run(() =>
-                {
-                    if (e.ImageFormat == SignatureImageFormat.Png)
-                    {
-                        return image.AsPNG().AsStream();
-                    }
-                    if (e.ImageFormat == SignatureImageFormat.Jpg)
-                    {
-                        return image.AsJPEG().AsStream();
-                    }
-                    return null;
-                });
-            }
-        }
-
-        private void OnIsBlankRequested(object sender, SignaturePadView.IsBlankRequestedEventArgs e)
-        {
-            var ctrl = Control;
-            if (ctrl != null)
-            {
-                e.IsBlank = ctrl.IsBlank;
-            }
-        }
-
-        private void OnPointsRequested(object sender, SignaturePadView.PointsEventArgs e)
-        {
-            var ctrl = Control;
-            if (ctrl != null)
-            {
-                e.Points = ctrl.Points.Select(p => new Point(p.X, p.Y));
-            }
-        }
-
-        private void OnPointsSpecified(object sender, SignaturePadView.PointsEventArgs e)
-        {
-            var ctrl = Control;
-            if (ctrl != null)
-            {
-                ctrl.LoadPoints(e.Points.Select(p => new NativePoint((float)p.X, (float)p.Y)).ToArray());
-            }
-        }
-
-        /// <summary>
-        /// Update all the properties on the native view.
-        /// </summary>
-        private void UpdateAll()
-        {
-            if (Control == null || Element == null)
-            {
-                return;
-            }
-
-            if (Element.BackgroundColor != Color.Default)
-            {
-                Control.BackgroundColor = Element.BackgroundColor.ToUIColor();
-            }
-            if (!string.IsNullOrEmpty(Element.CaptionText))
-            {
-                Control.CaptionText = Element.CaptionText;
-            }
-            if (Element.CaptionTextColor != Color.Default)
-            {
-                Control.Caption.TextColor = Element.CaptionTextColor.ToUIColor();
-            }
-            if (!string.IsNullOrEmpty(Element.ClearText))
-            {
-                Control.ClearLabelText = Element.ClearText;
-            }
-            if (Element.ClearTextColor != Color.Default)
-            {
-                Control.ClearLabel.SetTitleColor(Element.ClearTextColor.ToUIColor(), UIControlState.Normal);
-            }
-            if (!string.IsNullOrEmpty(Element.PromptText))
-            {
-                Control.SignaturePromptText = Element.PromptText;
-            }
-            if (Element.PromptTextColor != Color.Default)
-            {
-                Control.SignaturePrompt.TextColor = Element.PromptTextColor.ToUIColor();
-            }
-            if (Element.SignatureLineColor != Color.Default)
-            {
-                Control.SignatureLineColor = Element.SignatureLineColor.ToUIColor();
-            }
-            if (Element.StrokeColor != Color.Default)
-            {
-                Control.StrokeColor = Element.StrokeColor.ToUIColor();
-            }
-            if (Element.StrokeWidth > 0)
-            {
-                Control.StrokeWidth = Element.StrokeWidth;
-            }
-        }
-
-        /// <summary>
-        /// Update a specific property on the native view.
-        /// </summary>
-        private void Update(string property)
-        {
-            if (Control == null || Element == null)
-            {
-                return;
-            }
-
-            if (property == SignaturePadView.BackgroundColorProperty.PropertyName)
-            {
-                Control.BackgroundColor = Element.BackgroundColor.ToUIColor();
-            }
-            else if (property == SignaturePadView.CaptionTextProperty.PropertyName)
-            {
-                Control.CaptionText = Element.CaptionText;
-            }
-            else if (property == SignaturePadView.CaptionTextColorProperty.PropertyName)
-            {
-                Control.Caption.TextColor = Element.CaptionTextColor.ToUIColor();
-            }
-            else if (property == SignaturePadView.ClearTextProperty.PropertyName)
-            {
-                Control.ClearLabelText = Element.ClearText;
-            }
-            else if (property == SignaturePadView.ClearTextColorProperty.PropertyName)
-            {
-                Control.ClearLabel.SetTitleColor(Element.ClearTextColor.ToUIColor(), UIControlState.Normal);
-            }
-            else if (property == SignaturePadView.PromptTextProperty.PropertyName)
-            {
-                Control.SignaturePromptText = Element.PromptText;
-            }
-            else if (property == SignaturePadView.PromptTextColorProperty.PropertyName)
-            {
-                Control.SignaturePrompt.TextColor = Element.PromptTextColor.ToUIColor();
-            }
-            else if (property == SignaturePadView.SignatureLineColorProperty.PropertyName)
-            {
-                Control.SignatureLineColor = Element.SignatureLineColor.ToUIColor();
-            }
-            else if (property == SignaturePadView.StrokeColorProperty.PropertyName)
-            {
-                Control.StrokeColor = Element.StrokeColor.ToUIColor();
-            }
-            else if (property == SignaturePadView.StrokeWidthProperty.PropertyName)
-            {
-                Control.StrokeWidth = Element.StrokeWidth;
-            }
-        }
-    }
-}
